Split app data storage modules
